# openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Blogware RESTful API",
...
        "version": "1.1.2",
        "contact": {
            "name": "Blogware Support",
            "email": "alanmoehammad@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "x-logo": {
            "url": "https://blogware.site/public/files/pictures/scriptlog-1200x630.jpg"
        }
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Relative to the host serving this specification. The spec itself is served from /api/v1/openapi.json, so this resolves correctly on any install with no URL substitution needed. Human-readable examples elsewhere in this document use http://blogware.site as the example host."
        }
    ],
...
    "paths": {
...
        "/comments": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "List all approved comments",
                "description": "Retrieves a paginated list of approved comments.\nThis endpoint is public and does not require authentication.\n",
                "operationId": "listComments",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Page"
                    },
                    {
                        "$ref": "#/components/parameters/PerPage"
                    },
                    {
                        "$ref": "#/components/parameters/SortBy"
                    },
                    {
                        "$ref": "#/components/parameters/SortOrder"
                    },
                    {
                        "name": "post_id",
                        "in": "query",
                        "description": "Filter comments by post ID",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comments retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommentListResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            },
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Create a new comment",
                "description": "Creates a new comment on a blog post.\nThis endpoint is public and does not require authentication.\nComments are submitted with 'pending' status for moderation.\n",
                "operationId": "createComment",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CommentCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Comment created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommentResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/UnprocessableEntity"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "X-API-Key",
                "description": "API key authentication. Pass your API key in the X-API-Key header.\n"
            },
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Bearer token authentication. Pass your bearer token in the Authorization header.\n"
            }
        },
        "parameters": {
            "Page": {
                "name": "page",
                "in": "query",
                "description": "Page number for pagination",
                "required": false,
                "schema": {
                    "type": "integer",
                    "minimum": 1,
                    "default": 1,
                    "example": 1
                }
            },
            "PerPage": {
                "name": "per_page",
                "in": "query",
                "description": "Number of items per page",
                "required": false,
                "schema": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 100,
                    "default": 10,
                    "example": 10
                }
            },
            "SortBy": {
                "name": "sort_by",
                "in": "query",
                "description": "Field to sort by",
                "required": false,
                "schema": {
                    "type": "string",
                    "default": "ID",
                    "enum": [
                        "ID",
                        "post_date",
                        "post_modified",
                        "post_title",
                        "topic_title",
                        "comment_date"
                    ]
                }
            },
            "SortOrder": {
                "name": "sort_order",
                "in": "query",
                "description": "Sort direction",
                "required": false,
                "schema": {
                    "type": "string",
                    "enum": [
                        "ASC",
                        "DESC"
                    ],
                    "default": "DESC"
                }
            },
...
        },
        "schemas": {
            "Error": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "status": {
                        "type": "integer",
                        "example": 400
                    },
                    "error": {
                        "type": "object",
                        "properties": {
                            "code": {
                                "type": "string",
                                "example": "BAD_REQUEST"
                            },
                            "message": {
                                "type": "string",
                                "example": "Error description"
                            },
                            "details": {
                                "type": "array",
                                "items": {
                                    "type": "object",
                                    "properties": {
                                        "field": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "SuccessResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "status": {
                        "type": "integer",
                        "example": 200
                    },
                    "message": {
                        "type": "string",
                        "example": "Operation successful"
                    },
                    "data": {
                        "type": "object"
                    },
                    "_links": {
                        "$ref": "#/components/schemas/Links"
                    }
                }
            },
...
            "Comment": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "post_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "parent_id": {
                        "type": "integer",
                        "example": 0
                    },
                    "author": {
                        "type": "object",
                        "properties": {
                            "name": {
                                "type": "string",
                                "example": "John Doe"
                            },
                            "email": {
                                "type": "string",
                                "example": "john@example.com"
                            }
                        }
                    },
                    "content": {
                        "type": "string",
                        "example": "Great article!"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "approved",
                            "pending",
                            "spam"
                        ],
                        "example": "approved"
                    },
                    "date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15 10:30:00"
                    }
                }
            },
            "CommentCreate": {
                "type": "object",
                "required": [
                    "comment_author_name",
                    "comment_content",
                    "comment_post_id"
                ],
                "properties": {
                    "comment_author_name": {
                        "type": "string",
                        "maxLength": 60,
                        "example": "John Doe"
                    },
                    "comment_author_email": {
                        "type": "string",
                        "format": "email",
                        "nullable": true,
                        "example": "john@example.com"
                    },
                    "comment_content": {
                        "type": "string",
                        "example": "Great article!"
                    },
                    "comment_post_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "comment_parent_id": {
                        "type": "integer",
                        "default": 0,
                        "example": 0
                    }
                }
            },
...
            "CommentListResponse": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                        "type": "object",
                        "properties": {
                            "data": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Comment"
                                }
                            },
                            "pagination": {
                                "$ref": "#/components/schemas/Pagination"
                            }
                        }
                    }
                ]
            },
            "CommentResponse": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                        "type": "object",
                        "properties": {
                            "data": {
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/Comment"
                                    },
                                    {
                                        "type": "object",
                                        "properties": {
                                            "post": {
                                                "type": "object",
                                                "properties": {
                                                    "title": {
                                                        "type": "string"
                                                    },
                                                    "slug": {
                                                        "type": "string"
                                                    },
                                                    "url": {
                                                        "type": "string"
                                                    }
                                                },
                                                "nullable": true
                                            }
                                        }
                                    }
                                ]
                            }
                        }
                    }
                ]
            },
...
            "Link": {
                "type": "object",
                "description": "HATEOAS link following RFC 5988 (Web Linking)",
                "properties": {
                    "href": {
                        "type": "string",
                        "description": "The URL of the linked resource",
                        "example": "http://blogware.site/api/v1/posts"
                    },
                    "rel": {
                        "type": "string",
                        "description": "The relationship type",
                        "example": "self"
                    },
                    "type": {
                        "type": "string",
                        "description": "The HTTP method or MIME type",
                        "example": "GET"
                    },
                    "templated": {
                        "type": "boolean",
                        "description": "Whether the URL is a URI template",
                        "example": false
                    }
                },
                "required": [
                    "href",
                    "rel"
                ]
            },
            "Links": {
                "type": "object",
                "description": "HATEOAS links for discoverable API navigation",
                "additionalProperties": {
                    "$ref": "#/components/schemas/Link"
                },
                "example": {
                    "self": {
                        "href": "http://blogware.site/api/v1/posts?page=1",
                        "rel": "self",
                        "type": "GET"
                    },
                    "next": {
                        "href": "http://blogware.site/api/v1/posts?page=2",
                        "rel": "next",
                        "type": "GET"
                    },
                    "collection": {
                        "href": "http://blogware.site/api/v1/posts",
                        "rel": "collection",
                        "type": "GET"
                    }
                }
            },
            "Pagination": {
                "type": "object",
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "total_items": {
                        "type": "integer",
                        "example": 50
                    },
                    "total_pages": {
                        "type": "integer",
                        "example": 5
                    },
                    "has_next_page": {
                        "type": "boolean",
                        "example": true
                    },
                    "has_previous_page": {
                        "type": "boolean",
                        "example": false
                    }
                }
            },
...
        },
        "responses": {
            "BadRequest": {
                "description": "Bad Request - Invalid parameters",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
...
            "NotFound": {
                "description": "Not Found - Resource does not exist",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
...
            "UnprocessableEntity": {
                "description": "Unprocessable Entity - Validation failed",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
            "InternalServerError": {
                "description": "Internal Server Error",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
...
        }
    },
...
}